backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
django-environ==0.11.2
django-filter==23.3
djangorestframework==3.14.0
environ==1.0
greenlet==3.0.1
Markdown==3.5
msgpack==1.0.7
pynvim==0.4.3
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0

